requirements.txt
alembic==0.7.5.post2
Flask==0.10.1
Flask-Migrate==1.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
psycopg2==2.6
python-chess==0.8.0
SQLAlchemy==1.0.0
Werkzeug==0.10.4
